[BugFix] Fix flush concurrency problem in hash join (#27081)
