+++ be/src/exec/spill/input_stream.cpp
@@ -114,9 +114,11 @@ public:
 private:
     size_t read_idx{};
     std::vector<ChunkPtr> _chunks;
+    DECLARE_RACE_DETECTOR(detect_get_next)
 };
 
 StatusOr<ChunkUniquePtr> RawChunkInputStream::get_next(SerdeContext& context) {
+    RACE_DETECT(detect_get_next, var1);
     if (read_idx >= _chunks.size()) {
         return Status::EndOfFile("eos");
     }
