+++ be/src/exec/spill/spiller.hpp
@@ -256,170 +256,6 @@ Status PartitionedSpillerWriter::flush_if_full(RuntimeState* state, TaskExecutor
     return Status::OK();
 }
 
-template <class ChunkProvider>
-Status PartitionedSpillerWriter::spill_partition(SerdeContext& ctx, SpilledPartition* partition,
-                                                 ChunkProvider&& provider) {
-    auto& serde = _spiller->serde();
-
-    if (partition->spill_writer->block() == nullptr) {
-        spill::AcquireBlockOptions opts;
-        opts.query_id = _runtime_state->query_id();
-        opts.plan_node_id = options().plan_node_id;
-        opts.name = options().name;
-        ASSIGN_OR_RETURN(auto block, _spiller->block_manager()->acquire_block(opts));
-        std::lock_guard<std::mutex> l(_mutex);
-        partition->spill_writer->block_group().append(block);
-        partition->spill_writer->block() = block;
-    }
-
-    auto& block = partition->spill_writer->block();
-    DCHECK(block != nullptr);
-
-    auto consumer = [&](const auto& chunk) {
-        RETURN_IF_ERROR(serde->serialize(ctx, chunk, block));
-        partition->bytes += chunk->memory_usage();
-        return Status::OK();
-    };
-
-    RETURN_IF_ERROR(provider(consumer));
-
-    if (partition->spill_writer->block()->size() > options().spill_mem_table_bytes_size) {
-        RETURN_IF_ERROR(block->flush());
-        RETURN_IF_ERROR(_spiller->block_manager()->release_block(block));
-        block.reset();
-    }
-    return Status::OK();
-}
-
-template <class Consumer>
-class AccumulateWriter {
-public:
-    AccumulateWriter(Consumer& consumer, size_t chunk_size) : _consumer(consumer) {
-        _accumulator.set_max_size(chunk_size);
-    }
-
-    Status write(const ChunkPtr& chunk) {
-        DCHECK(_accumulator.need_input());
-        _accumulator.push(chunk);
-        if (_accumulator.has_output()) {
-            auto normalized = std::move(_accumulator.pull());
-            RETURN_IF_ERROR(_consumer(normalized));
-        }
-        return Status::OK();
-    }
-
-    Status flush() {
-        _accumulator.finalize();
-        if (auto chunk = std::move(_accumulator.pull())) {
-            if (!chunk->is_empty()) {
-                RETURN_IF_ERROR(_consumer(chunk));
-            }
-        }
-        return Status::OK();
-    }
-
-private:
-    Consumer& _consumer;
-    ChunkPipelineAccumulator _accumulator;
-};
-
-struct DoSpillPartition {
-public:
-    DoSpillPartition(SerdeContext& spill_ctx_, SpilledPartition* partition_, PartitionedSpillerWriter* writer_)
-            : spill_ctx(spill_ctx_), partition(partition_), writer(writer_) {}
-
-    Status operator()(const ChunkPtr& chunk) {
-        return writer->spill_partition(spill_ctx, partition, [&](auto& consumer) {
-            RETURN_IF_ERROR(consumer(chunk));
-            partition->num_rows += chunk->num_rows();
-            return Status::OK();
-        });
-    }
-
-private:
-    SerdeContext& spill_ctx;
-    SpilledPartition* partition;
-    PartitionedSpillerWriter* writer;
-};
-
-template <class MemGuard>
-Status PartitionedSpillerWriter::_split_partition(SerdeContext& spill_ctx, SpillerReader* reader,
-                                                  SpilledPartition* partition, SpilledPartition* left_partition,
-                                                  SpilledPartition* right_partition, MemGuard& guard) {
-    size_t current_level = partition->level;
-    size_t restore_rows = 0;
-
-    DoSpillPartition spill_left_partition(spill_ctx, left_partition, this);
-    DoSpillPartition spill_right_partition(spill_ctx, right_partition, this);
-
-    AccumulateWriter left_accumulate_writer(spill_left_partition, _runtime_state->chunk_size());
-    AccumulateWriter right_accumulate_writer(spill_right_partition, _runtime_state->chunk_size());
-
-    auto defer = DeferOp([&]() {
-        left_accumulate_writer.flush();
-        right_accumulate_writer.flush();
-    });
-
-    while (true) {
-        RETURN_IF_ERROR(reader->trigger_restore(_runtime_state, SyncTaskExecutor{}, guard));
-        if (!reader->has_output_data()) {
-            DCHECK_EQ(restore_rows, partition->num_rows);
-            break;
-        }
-        ASSIGN_OR_RETURN(auto chunk, reader->restore(_runtime_state, SyncTaskExecutor{}, guard));
-        restore_rows += chunk->num_rows();
-        if (chunk->is_empty()) {
-            continue;
-        }
-        auto hash_column = down_cast<SpillHashColumn*>(chunk->columns().back().get());
-        const auto& hash_data = hash_column->get_data();
-        // hash data
-        std::vector<uint32_t> shuffle_result;
-        shuffle_result.resize(hash_data.size());
-        size_t left_channel_size = 0;
-        for (size_t i = 0; i < hash_data.size(); ++i) {
-            shuffle_result[i] = hash_data[i] >> current_level & 0x01;
-            left_channel_size += !shuffle_result[i];
-        }
-        size_t left_cursor = 0;
-        size_t right_cursor = left_channel_size;
-        std::vector<uint32_t> selection(hash_data.size());
-        for (size_t i = 0; i < hash_data.size(); ++i) {
-            if (shuffle_result[i] == 0) {
-                selection[left_cursor++] = i;
-            } else {
-                selection[right_cursor++] = i;
-            }
-        }
-
-#ifndef NDEBUG
-        for (size_t i = 0; i < left_cursor; i++) {
-            DCHECK_EQ(hash_data[selection[i]] & left_partition->mask(),
-                      left_partition->partition_id & left_partition->mask());
-        }
-
-        for (size_t i = left_cursor; i < right_cursor; i++) {
-            DCHECK_EQ(hash_data[selection[i]] & right_partition->mask(),
-                      right_partition->partition_id & right_partition->mask());
-        }
-#endif
-
-        if (left_channel_size > 0) {
-            ChunkPtr left_chunk = chunk->clone_empty();
-            left_chunk->append_selective(*chunk, selection.data(), 0, left_channel_size);
-            left_accumulate_writer.write(left_chunk);
-        }
-        if (hash_data.size() != left_channel_size) {
-            ChunkPtr right_chunk = chunk->clone_empty();
-            right_chunk->append_selective(*chunk, selection.data(), left_channel_size,
-                                          hash_data.size() - left_channel_size);
-            right_accumulate_writer.write(right_chunk);
-        }
-    }
-    DCHECK_EQ(restore_rows, partition->num_rows);
-    return Status::OK();
-}
-
 template <class TaskExecutor, class MemGuard>
 Status PartitionedSpillerWriter::flush(RuntimeState* state, bool is_final_flush, TaskExecutor&& executor,
                                        MemGuard&& guard) {
@@ -430,70 +266,29 @@ Status PartitionedSpillerWriter::flush(RuntimeState* state, bool is_final_flush,
         return Status::OK();
     }
 
+    if (is_final_flush && _running_flush_tasks > 0) {
+        _need_final_flush = true;
+        return Status::OK();
+    }
+    DCHECK_EQ(_running_flush_tasks, 0);
     _running_flush_tasks++;
 
     auto task = [this, state, guard = guard, splitting_partitions = std::move(splitting_partitions),
                  spilling_partitions = std::move(spilling_partitions), trace = TraceInfo(state)]() {
         SCOPED_SET_TRACE_INFO({}, trace.query_id, trace.fragment_id);
         RETURN_IF(!guard.scoped_begin(), Status::Cancelled("cancelled"));
-        DCHECK_EQ(_running_flush_tasks, 1);
+        RACE_DETECT(detect_flush, var1);
+        // concurrency test
         auto defer = DeferOp([&]() {
             _spiller->update_spilled_task_status(_decrease_running_flush_tasks());
             guard.scoped_end();
         });
 
-        // partition memory usage
-        // now we partitioned sorted spill
-        SerdeContext spill_ctx;
-        {
-            SCOPED_TIMER(_spiller->metrics().flush_timer);
-            for (auto partition : spilling_partitions) {
-                RETURN_IF_ERROR(spill_partition(spill_ctx, partition, [&partition](auto& consumer) {
-                    auto& mem_table = partition->spill_writer->mem_table();
-                    RETURN_IF_ERROR(mem_table->flush(consumer));
-                    return Status::OK();
-                }));
-            }
+        if (_spiller->is_cancel() || !_spiller->task_status().ok()) {
+            return Status::OK();
         }
 
-        {
-            SCOPED_TIMER(_spiller->metrics().split_partition_timer);
-            for (auto partition : splitting_partitions) {
-                auto [left, right] = partition->split();
-                left->spill_writer = std::make_unique<RawSpillerWriter>(_spiller, _runtime_state, _mem_tracker.get());
-                left->in_mem = false;
-                RETURN_IF_ERROR(left->spill_writer->prepare(state));
-                left->spill_writer->acquire_mem_table();
-                right->in_mem = false;
-
-                right->spill_writer = std::make_unique<RawSpillerWriter>(_spiller, _runtime_state, _mem_tracker.get());
-                RETURN_IF_ERROR(right->spill_writer->prepare(state));
-                right->spill_writer->acquire_mem_table();
-
-                // write
-                std::shared_ptr<SpillInputStream> stream;
-                RETURN_IF_ERROR(partition->spill_writer->acquire_stream(&stream));
-
-                auto reader = std::make_unique<SpillerReader>(_spiller);
-                reader->set_stream(std::move(stream));
-
-                // split process may be generate many small chunks. we should fix it
-                auto st = _split_partition(spill_ctx, reader.get(), partition, left.get(), right.get(), guard);
-                DCHECK(st.ok() || st.is_end_of_file());
-                DCHECK_EQ(left->num_rows + right->num_rows, partition->num_rows);
-
-                left->spill_writer->acquire_mem_table();
-                right->spill_writer->acquire_mem_table();
-
-                _add_partition(std::move(right));
-                _add_partition(std::move(left));
-            }
-
-            for (auto partition : splitting_partitions) {
-                _remove_partition(partition);
-            }
-        }
-
+        _spiller->update_spilled_task_status(_flush_task(splitting_partitions, spilling_partitions));
         return Status::OK();
     };
